First version
Here is our attempt at the XLST translator, main chunk of the work was
by Darren, it's a console application that applies an xsl (i.e.
circle.xsl) to some xml (i.e. number.xml) and writes out the result to a
console window.

The translation is based upon our database resource translator i.e. a
class that creates a resource manager based on the default/neutral
resource resx, and implements a Find method. This method takes a string
that we want to find (specified in the default language i.e. in our case
English) which is then effectively used as a key into the resource
file(s). The text that is passed into the Find method is first converted
into a valid identifier (e.g. spaces removed, slashes removed etc...
etc...), before it can be used as a key. Then, if a key exists for the
specified culture (French in the case of the test solution) then the
appropriate language version of the original passed in text is returned.

In order to use the resource translator in the xsl, an object of this
type is first created and then passed into the xsl using extension
objects. The Find method can then be called by referencing the correct
extension object, and specifying the piece of text that we would like to
be translated.